[Silver III] Title: 주유소, Time: 20 ms, Memory: 2804 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -8,7 +8,7 @@ vector<int> dist;
 void solve() {
     long long ans = 0;
     for (int i = 0; i < N - 1; ++i) {
-        ans += price[i] * dist[i];
+        ans += (long long)price[i] * dist[i];
     }
     cout << ans;
 }
